airflow/requirements.txt
pandas==2.2.2
numpy==1.23.5
matplotlib==3.8.4
scikit-learn==1.4.2
kafka-python==2.0.2
mlflow==2.12.1
tensorflow==2.15.1
azure-identity==1.16.0
azure-storage-blob==12.19.1
azureml-core==1.56.0
stockstats==0.6.2